+++ packages/strava_api/lib/src/models/detailed_activity.dart
@@ -58,9 +58,9 @@ class DetailedActivity {
   //@JsonKey(defaultValue: false)
   //@JsonKey(required: true)
   //@JsonKey(ignore: true)
-  final BigInt? id;
+  final int? id;
   final String? externalId;
-  final BigInt? uploadId;
+  final int? uploadId;
   final MetaAthlete? athlete;
   final String? name;
   final double? distance;
